Finished lights logic.
Finished logic on the lights.  Using strings instead of char, ints, bytes, etc to hold leading zeros that are present in the binary numbers due to issues with ambiguity.  Simply needs light patterns and additions to the if-statement chain to be done.  Using GOTO as a meme because I felt like it.